Optimize hashCodeOfString
