+++ nodejs/src/consumer/FilterExpression.ts
@@ -16,6 +16,7 @@
  */
 
 import { FilterType, FilterExpression as FilterExpressionPB } from '../../proto/apache/rocketmq/v2/definition_pb';
+import { hashCodeOfString } from '../util';
 
 const TAG_EXPRESSION_SUB_ALL = '*';
 
@@ -45,20 +46,8 @@ export class FilterExpression {
 
   hashCode(): number {
     let hash = 17;
-    hash = hash * 31 + this.hashCodeOfString(this.expression);
+    hash = hash * 31 + hashCodeOfString(this.expression);
     hash = hash * 31 + this.filterType;
-    return hash;
-  }
-
-  private hashCodeOfString(str: string): number {
-    let hash = 0;
-    for (let i = 0; i < str.length; i++) {
-      const char = str.charCodeAt(i);
-      // eslint-disable-next-line no-bitwise
-      hash = ((hash << 5) - hash) + char;
-      // eslint-disable-next-line no-bitwise
-      hash = hash & hash; // Convert to 32bit integer
-    }
     return hash;
   }
